requirements.txt
# Core Dependencies
aiogram==3.12.0              # Telegram Bot framework
APScheduler==3.10.4          # Task scheduling
python-dotenv==1.0.1         # Environment variables

# Web & HTTP
httpx>=0.27.0                # Async HTTP client
aiohttp==3.9.1               # Async HTTP framework
requests>=2.31.0             # HTTP library

# Parsing & Processing
feedparser==6.0.11           # RSS/Atom feed parsing
beautifulsoup4==4.12.3       # HTML parsing
lxml==5.3.0                  # XML/HTML parser

# AI & Machine Learning
groq==0.31.1                 # Groq API client for AI summarization

# Async & Reliability
aiofiles>=23.2.1             # Async file operations
tenacity>=8.2.3              # Retry logic for reliability
